chore: add prod ad ved mode
