+++ package.json
@@ -1,13 +1,11 @@
 {
   "name": "crud-api",
   "version": "1.0.0",
-  "main": "index.js",
   "type": "module",
   "scripts": {
-    "build:prod": "webpack --config webpack.prod.js",
-    "build:dev": "webpack --config webpack.dev.js",
-    "start:prod": "webpack serve --config webpack.prod.js",
-    "start:dev": "node --no-warnings=ExperimentalWarning --loader ts-node/esm src/index.ts"
+    "build": "webpack",
+    "start:prod": "webpack && node dist/bundle.cjs",
+    "start:dev": "node --loader ts-node/esm src/index.ts"
   },
   "keywords": [],
   "author": "",
@@ -18,6 +16,7 @@
     "@types/uuid": "^10.0.0",
     "dotenv": "^16.4.5",
     "nodemon": "^3.1.7",
+    "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.6.3",
     "uuid": "^10.0.0",
