+++ tsconfig.json
@@ -1,18 +1,19 @@
 {
   "compilerOptions": {
-    "target": "ESNext",
+    "typeRoots": ["./node_modules/@types", "./@types"],
+    "target": "ES2020",
     "module": "ESNext",
     "moduleResolution": "node",
     "esModuleInterop": true,
     "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
     "strict": true,
-    "allowImportingTsExtensions": true,
-    "noEmit": true
+    "noEmit": false,
+    "rootDir": "./src"
   },
   "ts-node": {
-    "esm": true
+    "esm": true,
+    "experimentalSpecifierResolution": "node"
   },
-  "include": ["src/**/*.ts", "src/index.mjs", "src/index.ts"],
+  "include": ["src/**/*.ts", "@types/**/*.d.ts"],
   "exclude": ["node_modules"]
 }
