--- webpack.common.js
@@ -1,32 +1,0 @@
-import { fileURLToPath } from 'url';
-import { dirname, resolve as _resolve } from 'path';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
-const commonConfig = {
-  entry: './src/index.ts',
-  target: 'node',
-  module: {
-    rules: [
-      {
-        test: /\.ts$/,
-        use: 'ts-loader',
-        exclude: /node_modules/,
-      },
-    ],
-  },
-  resolve: {
-    extensions: ['.ts', '.js'],
-  },
-  output: {
-    path: _resolve(__dirname, 'dist'),
-    filename: 'bundle.js',
-    libraryTarget: 'commonjs2',
-  },
-  experiments: {
-    outputModule: true,
-  },
-};
-
-export default commonConfig;
