+++ webpack.config.js
@@ -1,0 +1,35 @@
+import { resolve } from 'node:path';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const config = {
+  mode: 'production',
+  entry: './src/index.ts',
+  output: {
+    filename: 'bundle.cjs',
+    path: resolve(__dirname, 'dist'),
+  },
+  resolve: {
+    extensions: ['.ts', '.js'],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        use: 'ts-loader',
+        exclude: /node_modules/,
+      },
+    ],
+  },
+  target: 'node',
+  devServer: {
+    static: './dist/bundle.сjs',
+    port: 3000,
+    hot: true,
+  },
+};
+
+export default config;
